LayoutTests/platform/ios-sim-deprecated/media/audio-mpeg4-supported-expected.txt

Test that the audio element supports M4A files.

EVENT(canplaythrough)
EXPECTED (mediaElement.currentTime == '0') OK
EVENT(canplaythrough)
EXPECTED (mediaElement.currentTime == '0') OK
EVENT(play)
EXPECTED (mediaElement.currentTime > '0') OK
END OF TEST
